move snapshotter creator to rest
