+++ adapters/handlers/rest/configure_api.go
@@ -32,7 +32,6 @@ import (
 	"github.com/semi-technologies/weaviate/adapters/handlers/rest/clusterapi"
 	"github.com/semi-technologies/weaviate/adapters/handlers/rest/operations"
 	"github.com/semi-technologies/weaviate/adapters/handlers/rest/state"
-	"github.com/semi-technologies/weaviate/adapters/repos/backups"
 	"github.com/semi-technologies/weaviate/adapters/repos/classifications"
 	"github.com/semi-technologies/weaviate/adapters/repos/db"
 	"github.com/semi-technologies/weaviate/adapters/repos/db/inverted"
@@ -188,7 +187,7 @@ func configureAPI(api *operations.WeaviateAPI) http.Handler {
 	shardingStateFunc := func(className string) *sharding.State {
 		return appState.SchemaManager.ShardingState(className)
 	}
-	snapshotterProvider := backups.NewSnapshotterProvider(repo)
+	snapshotterProvider := newSnapshotterProvider(repo)
 	backupManager := backup.NewBackupManager(appState.Logger, snapshotterProvider, appState.Modules, shardingStateFunc)
 
 	// TODO: configure http transport for efficient intra-cluster comm
